+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Fix jobs loading issue when Routemaster::Jobs is not defined while executing
 `Kernel.const_get`.
 - Make response caching configurable. Make index request to not cache the responses with the urls
+- Support Distributed Redis configuration for load balancing
 
 ### 2.3.0 (2017-01-16)